Some optimizations

+++ BackendHomework.Core/Services/PlateService.cs
@@ -1,9 +1,7 @@
-﻿using BackendHomework.Core.DTOs;
-using BackendHomework.Core.Entities;
+﻿using BackendHomework.Core.Entities;
 using BackendHomework.Core.Exceptions;
 using BackendHomework.Core.Interfaces;
 using Microsoft.EntityFrameworkCore;
-using Newtonsoft.Json;
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
@@ -17,9 +15,23 @@ namespace BackendHomework.Core.Services
         {
             _plateRepository = plateRepository;
         }
-        public async Task<bool> DeletePlate(Plate plate)
+        public async Task<bool> DeletePlate(Guid plateId, string userId)
         {
-            return await _plateRepository.Delete(plate);
+            var plate = await this.GetPlate(plateId);
+
+            if (plate != null)
+            {
+                if (plate.UserId != userId)
+                {
+                    throw new BusinessException("The plate you are trying to delete does not belong to you, please try with another plate");
+                }
+
+                return await _plateRepository.Delete(plate);
+            }
+            else
+            {
+                throw new BusinessException("The plate you are trying to delete does not exist, please try with another plate");
+            }
         }
 
         public async Task<bool> DeleteAllUserPlates(string userId) 
@@ -70,18 +82,21 @@ namespace BackendHomework.Core.Services
 
         public async Task<bool> UpdatePlate(Plate plate,string userId)
         {
-            var plateOld =  this.GetPlate(plate.Id).Result;
+            var plateOld =  await this.GetPlate(plate.Id);
+
             if (plateOld != null)
             {
+                if (plateOld.UserId != userId)
+                {
+                    throw new BusinessException("The plate you are trying to edit does not belong to you, please try with another plate");
+                }
+
                 plateOld.Description = plate.Description;
                 plateOld.Price = plate.Price;
                 plateOld.Name = plate.Name;
                 plateOld.Type = plate.Type;
-                if (plate.UserId != userId)
-                {
-                    throw new BusinessException("The plate you are trying to edit does not belong to you, please try with another plate");
-                }
-                return await _plateRepository.Update(plate);
+                
+                return await _plateRepository.Update(plateOld);
             }
             else
             {
